requirements-infer.txt
fastapi
uvicorn
python-multipart

torch
torchaudio
pytorch-lightning

numpy
soundfile
librosa
scipy
scikit-learn
matplotlib

pyyaml

